DjangoREST/Projeto_Escola/requirements.txt
asgiref==3.7.2
Django==3.0.7
djangorestframework==3.11.0
Markdown==3.2.2
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
